applied bigdecimal patch

+++ src/main/java/pizza/order/Order.java
@@ -2,6 +2,7 @@ package pizza.order;
 
 import pizza.customer.Customer;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class Order {
@@ -14,7 +15,7 @@ public class Order {
 
     private final Customer customer;
 
-    private final Double totalPrice;
+    private final BigDecimal totalPrice;
 
     private final LocalDateTime estimatedTimeOfDelivery;
 
@@ -22,13 +23,13 @@ public class Order {
     // --- constructors ---
     //
 
-    public Order(Customer customer, Double totalPrice, LocalDateTime estimatedTimeOfDelivery) {
+    public Order(Customer customer, BigDecimal totalPrice, LocalDateTime estimatedTimeOfDelivery) {
         this.customer = customer;
         this.totalPrice = totalPrice;
         this.estimatedTimeOfDelivery = estimatedTimeOfDelivery;
     }
 
-    public Order(Long id, Customer customer, Double totalPrice, LocalDateTime estimatedTimeOfDelivery) {
+    public Order(Long id, Customer customer, BigDecimal totalPrice, LocalDateTime estimatedTimeOfDelivery) {
         this.id = id;
         this.customer = customer;
         this.totalPrice = totalPrice;
@@ -54,7 +55,7 @@ public class Order {
         return customer;
     }
 
-    public Double getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
